Implemented PerlDL interface for the low-level fucntion plshade1.  This allows the example x15c.c in the PLplot sources to be fully ported to PerlDL.
With this addition, I generalized the *_callback/*_subroutine approach for
passing subroutine names/references to the PerlDL/PLplot API functions.
The plshade1 function needs three such subroutine arguments: defined, fill,
and pltr.  For now, the fill argument is always passed as plfill to
c_plshade1.  Inn part, the reason for doing so is that plfill is effectively
the only fill routine available in PLplot.  Besides that, using a subroutine
which accepts piddle arguments (as the PerlDL plfill) would complicate the
code in pp_def('plshade1').  This may change in the future.