Changed VERSION to 2.4.4_05git
The $PDL::VERSION is changed to 2.4.4_05git for more
development.  The pattern I've decided to try is the
development release for a given version, say 2.4.4_05
is the same base with the string 'git' added on the
end.  When the release is made, the git suffix is
removed and the distribution gnerated from there.