Fix readflex() to work with File::LinearRaid
readflex() in IO/FlexRaw/FlexRaw.pm use method syntax for
read with an input file handle.  This failed with a tied
file handle as is used by File::LinearRaid.  Replacing the
usage $d->read(...) by read($d, ...) fixed the incompatibility.
I don't know if it should have worked the other way or not.