Update version to 2.4.4_12_1 for development
With the rapid releases, we need a unique identifier for
each git push to try.  See if adding one more digit rather
than git works.  The plan is to update the last digit for
each push to sf.net so there is a version "paper trail"
for reference and debugging.