Change PP typemap finding code again, ensure Makefile.PL honours USE_POGL setting
The new PP typemap finding logic was broken. Realised this on OS X after upgrading MakeMaker, which installs the new MakeMaker in an update directory. There is no 'typemap' file in that update dir, BOOM... PP falls flat on its face. The latest attempt is based on what I found in MM_Unix.pm, although way less general in terms of checking for special cases. If this does not work portably, we should consult a site like Perl Monks or similar and split this code out.

The Open_GL test in toplevel Makefile.PL did not honour the USE_POGL config setting. I thought it should.