Partial fix to PDL::IO::Browser
Only tested with cygwin/XP.  The Makefile.PL logic to determine
if curses/ncurses are present and to use them needs to be verified
and extended to support win32 (e.g. if PDcurses is installed).

There appears to be functionality relating to setting values in
the piddle being viewed.  This needs to be cleaned up and
documented.  As is, it can get pretty confusing.  Also need
to add redraw (via Ctrl-L), and Page Up, Page Down, etc.