POGL_WINDOW_TYPE now sets OpenGL window type
If you have the the needed X11 bindings
in your POGL then you can use the
$PDL::Config{POGL_WINDOW_TYPE} to select
which flavor of OpenGL window is created.

The X11 interface has been preserved to
enable the use of PDL::Graphics::TriD::Tk
which provides a Tk widget interface to
PDL::Graphics::TriD. The existing widget
implementation only works for X11.

It appears that Gtk+ may actually have a
supported OpenGL widget which might make
better sense for a full featured widget set
to support.