Adding a GLUT font example code for debugging
There have been multiple reports of GLUT failures on
linux machines with the symptom being font not found.
This is a C program that users can try to build to see
if the problem is in the bindings or in their local
FreeGLUT install.  I plan to move this to utils and
put in some sort of build information.