Make window the *GLUT* current window for display
This doesn't make any difference with only one window but
with multiple windows, the image was appearing in the
wrong windows.  Still need a better way to select/set an
active figure window. There are currently two globals for
the TriD current window:

  $PDL::Graphics::TriD::cur
  $PDL::Graphics::TriD::current_window

Need to determine if they are both needed and if not,
get rid of one of them.