Modify rint algorithm for non-GNU compilers.
On GNU compilers, PDL::Math::rint uses the system library function. But on Microsoft compilers, there is no rint. The rint that was provided handled half-integer rounding differently than the GNU implementation. This commit implements a simple Banker's rounding algorithm that can be used by Microsoft compilers, updates the docs for rint to describe the calculation being performed (as well as providing some additional rounding guidance), and adds a new test file for this function.