Rewrote cat's error processing so that it's faster and more informative.
The original cat code would barf, but the problem is that it would do so
with a retarded message.  I revised my first attempts at solving this by
putting the original cat code inside an eval.  After the eval, if there
was an error, I go on a big parsing spree to figure out what all the
user may have screwed up.