Refactor rcols to support memory-time perf trades
A profile of PDL::rcols shows that much/most of the
time is spent in the PDL set (and internal set_c) calls
to update the column piddles with the values as they
are read.  This refactors the basic data structures for
the columns to allow for associated perl arrays for all
columns.  Then we can call the expensive (in latency)
PDL set/assignment routines less often with the hope of
speeding up the overall IO performance.