Refactor _ext_1D to _ext_lastD for rcols
In order to support multiple columns in one piddle
with rcols, need to have the extend routine work on
the last dim of an arbitrary piddle and not just
dim 0 of a 1D piddle.

A cool thing about the implementation was that I
could use ->mv(-1,)->slice()->mv(0,-1) to use the
same, simple slice expression as in the original
code.  Way to go PDL!