Refactor data structure for rcols
This is a minor change of data structure used for
chunking the rcols IO.  Originally, this was [ X, a, b, c... ]
where X was either a PDL or ARRAY and a, b, c... were the
elements being collected.  It makes book keeping and access
much simpler if I use [ PDL, [] ] or [ [], [] ] since now
all the new elements can be directly accessed without shifts
or pops.  Conveniently, pdl() take either arrays or array
refs.