Fix failure to strip PERLCOLS in rcols
If a column is specified for input with rcols() but also
listed in PERLCOLS, then the position of the data in output
is as listed in the explicit column specification location
but the type is a perl array ref.  This fixes an oops where
I had failed to stop reading into the piddle.