Add hooks for PDL::new to handle STRING data
I did not have time to to a full implementation for the
new Matlab-style (comma free) string data option for the
pdl routine (a.k.a. PDL::new under the hood).  I did get
a basic dispatch for the new option working which uses
a routine PDL::Core::new_pdl_from_string to do its magic.

If you define that routine, you'll be able to play with
options for the STRING input handling for pdl / PDL::new.
Type help PDL::new or look at the PDL/Basic/Core/Core.pm
source for how it works and what a proper new_pdl_from_string
routine needs to do.  A free example is included!