Update Padre/README re _plugin_locator
This is needed to display the banner of shell capabilities
on startup.  Another issue is how to enable and disable the
various features.  If that is not directly possible, maybe
it is possible to mark features to load, then save state,
and start another shell with the desired plugins loaded in
place of the existing shell instance.