Move TODO to wiki page for review
The current TODO file is languishing because of all the dead
wood.  It has been pointed out in recent mailing list threads
that irrelevant and incomplete TODO lists give a poor impression
of PDL development to potential new contributers.  Also, with
all the new interest, we need a way to keep focus on priorities
to keep moving forward.