Fixed varargs/stdargs issues with barf in slices.pd.
Reported by shlomi (rindolf) on irc, slices.pd was not compiling because
it was giving a non-constant string to barf, which is croak, which
behaves a lot like prinft. gcc insists that the specification string
to printf and similar functions is a string literal. This fixes that
problem in slices.pd.

It appears that the original code couldn't decide if barf just took
a single string argument or a variable number of arguments, so they
used sprintf and put the message in a buffer string. I simply
copied the contents of the sprintf statement and gave it to
the barf function call.