Add Convert::UU to PDL requirements list
Yes, I know this is yet another dependency but it is an all perl
replacement for the uuencode/uudecode needed for PDL::IO::Dumper.
This should not affect performance as the faster system calls to
uuencode/uudecode will still be made if available.  However, now
PDL::IO::Dumper should work on every PDL platform.