Update Release_Notes for Perldl2 l and demo info
The 'l' command works for listing history.
The 'demo' command works partially.  You will need to quote
the demo name argument and at the end of the demo, it may
hang waiting for input.  On unix-type systems, Ctrl-D should
exit at that time.  For win32, you'll have to exit the shell
and restart for now.