Don't load INC and Keywords completion wo deps
Now check for required dependencies for CompletionDriver::INC
and CompletionDriver::Keywords (File::Next and B::Keywords,
respectively) before loading the plugins.  Still needs a
refactor to simplify and make available for all plugins as
a check.  Also updated Perldl2/TODO with this information
as well as the investigation results into the use of clear
to clear the screen for win32.