Fix jpegtopnm problem in proj_transform.t
The problem arises because the jpegtopnm routine
recently started giving a different output from
extracting the same input earth_image data file from
PDL::Transform::Cartography. That results in test failure
because the data does not match but not because the
PDL::Transform::Proj4 has an error.

The fix is in two parts. Modify the original
t/proj_transform.t to skip the tests if the image data
checksum does not match that for which the tests were
written. Second, a new t/proj_transform2.t test was written
using a synthetic image in place of the earth_image. This is
possible because the image data is only used to compare the
test results but it does not contain any explicit content
related to the transform beyond the FITS header information.

This should fix sf.net bug #3013751.