Add INT2PTR to PP usage for IV to ptr conversion
The PDL internal code has been fixed for a while but the PP.pod
was never updated with the correction.  Without this macro
you can have mysterious bad pointers resulting from a pointer
being converted to an integer and then to a pointer.  This
only works if int is long enough to hold a pointer value.  By
construction, IVs are always big enough to hold a pointer.