Fix t/opengl.t false fail with no X display
glutInit exits on failure.  This tries opening the X
display before calling glutInit just in case.  The
tests involving glutInit are skipped if the display
cannot be open.

The required version of OpenGL was updated to 0.63
in the perldl.conf file.  It should be available on
a CPAN mirror shortly.