Fixed a bug in mapflex code (though others remain).
The mapflex function created a piddle with something that looked like

  '$piddle = byte zeroes(1)'

This used to be interpreted as a piddle with contents of zero, but
with the new piddle-from-string functionality, it was tripping an
error. It looks like broken behavior had a default result that the
piddle-from-string functionality has effectively removed. So I
fixed that.