fix wmpeg() to use ffmpeg to generate the video
This resolves sf.net feature request #3010984.
Note that ffmpeg supports much more video file functionality
and formats than mpeg_encode.  I would like to see this as
the basis for a general read/write AVI capability in the
future.  This should work for cygwin and unix/linux/*bsd/macosx
but needs to be verified on win32.