Possible fix to PDL::IO::Dumper failures
There have been CPAN tester failures from t/dumper.t where the
problem was scalar piddles werer being dumped as the value of the
memory location *but* as a signed integer.  That resulted in a
synthesized variable named '$PDL-12345', for example, which is
not valid perl.  The problem appears to be some implicit stringify
to decimal in IO/Dumper.pm.  I think that has been fixed by using
sprintf and %u formats.