Make t/flexraw.t a "TODO" test
Trying to come up with a standard for using CPAN Testers
to help diagnose PDL bugs.  The Test::More TODO capability
allows one to have tests corresponding to development TODOs.
Unless the $TODO is defined, then the test acts as normal.

By default, $TODO is defined if SKIP_KNOWN_PROBLEMS is set
in either the environment or the perldl.conf file.  However,
it isn't useful to generate fail messages for bugs that are
not being worked on (e.g. without a developer assigned to
the ticket).

If no developer is assigned, the default is to set the
local $TODO always.  If the ticket has an assigned developer,
then the setting will obey the SKIP_KNOWN_PROBLEMS options:
on for official releases, off for development releases.

If you are an assigned developer and wish to have test FAILs
ignored for your TODO items, just edit the local $TODO line
and remove/comment out the test for SKIP_KNOWN_PROBLEMS.