Create inc/ folder for local module copies
I ran the use-devel-checklib command to create
a PDL/inc folder for local modules used by PDL.
It starts with a copy of Devel::CheckLib with
the package line munged so it will not be indexed
by PAUSE/CPAN indexers.

The intent is to use this local inc/ folder to
develop the needed Alien::XXX modules for PDL
portability.  We can munge the package lines so
that they are not indexed until they are stable
and ready for release.  The first step will be
to implement detection functionality.  The next
step would be to implement install functionality.