Fix inc/Devel/CheckLib ASPerl issues
There were problems with Devel::CheckLib and ActiveState Perl.

 (1) require Devel::CheckLib fails with an error about
     %Config::Confg being a read only value

 (2) backslashes in the windows paths were deleted making
     the resulting @flags incorrect and unusable.

This fix makes an explicit working copy of the needed %Config
entries in the _findcc() routine which should resolve the
compile failure.  It also sets the $keep option to the
quotewords() routine to true which appears to fix the bad
win32 paths in the option @flags.  The fixes will be pushed
upstream but are here in the local copy so things work.