Fix sf.net #3171702 re: missing podselect command
Added Pod::Parser and Pod::Select to the PDL prerequisites
(they are in perl5 Core modules since v5.6 so this should
not be a problem).  Then replaced call to podselect script
by calling the podselect() function in Pod::Select directly.

This should help with the many solaris FAILs on CPAN
testers and set the way to port PDL::Doc and others to
use the Core Pod modules rather than the PDL::Pod::*
snapshot with special fixes.