Minor POD fixes for Core.pm.PL
There were some glitches in the CPAN search.cpan.org on-line
docs so I fixed the outstanding items in Core.pm.PL in the
hopes that it will resolve the problem.  The only podchecker
error was:

  *** ERROR: unterminated T<...> at line 2480 in file Core.pm.PL

coming from a line line:

  print OUT<<'!NO!SUBS!';

where the OUT<< was apparently being interpreted as a OU
followed by a T< ... construct.