Add refresh.pdl to Example/PLplot
This is from the sf.net patches tracker #3024385,
for an "imag routine for PLplot".  Until/if it is
folded into the PDL::Graphics::PLplot module, it
seemed to make sense having it as an example.

Please try it out if you have PLplot and be sure
to update or add documentation or usage notes
with any issues that arise.  I am not able to test
the routines since no PLplot on cygwin yet...