Implement PDL::Constants module
This uses the constant pragma to implement
PI and E constants.  There has been discussion
on a number of other constant values that would
be of use but it seemed simpler to start with
the basics.  This addresses sf.net feature request
ticket #2787823, "add math constants to PDL: pi, e,.."