Add bad value support to PDL::IO::FlexRaw
Still need to add some tests but the idea is that after
the Dims in the FlexRaw header, on the same line, if
$pdl->badflag then the token 'badvalue' is added.  If
the $pdl->badvalue is not the $pdl->orig_badvalue then
that value is added afterwards.  This is a workaround
to the problem of uniquely saving and restoring a float
or double value from a string.  Until that code is
in, we just support defaults.  NOTE: need to add a
test (and possibly code) to handle the NaN as bad value
case...