Update cygwin/INSTALL re rebaseall
Discusion on the cygwin-xfree mailing list
suggests specifying a different (higher) base
address might alleviate some of the situations
when rebaseall fails to work.  This could be
a problem though for PDL/perl since there are
large numbers of DLLs which could lead to the
addresses trampling high memory stuff.