PLplot: user prefs for library and include file locations take precedence over defaults
The locations $PDL::Config{WHERE_PLPLOT_LIBS} and {WHERE_PLPLOT_INCLUDES} were at the end of the list of paths to search. I moved them to the beginning of the list so that if there is more than one installed on the system the desired library/includes can be specified.