Fix =begin html/=end html usage in PP.pod
It looks like an edit was only partially done but the
result was that only the HTML docs for PP had the headings
for the various pp_xxx routine descriptions.  This made
normal non-html output difficult to understand or search.

I just removed the extraneous =begin html/=end html pairs
so the pp_xxx function heading appear in all versions of
the documentation.