Add comment for way forward w code refs and nslice
Plan to rewrite the nslice() method to detect when
we have the CODE ref usage.  In that case, UNIVERSAL::nslice()
would be called which would construct the desired CODE ref
method call using the original pre-NiceSlice version of the
nslice() arguments which would be stashed by the source filter
during the original processing.

NOTE: This has potential to collide with another UNIVERSAL::nslice()
method declaration.