Add nslice_if_pdl variant of nslice
This routine is the same as nslice but it takes
an extra argument from the PDL::NiceSlice filter
process, a string of the original args before
processing.  The idea is to allow a code reference
to be called if the $self is not a piddle.

This works provided $self is a blessed object.
Then we could use UNIVERSAL::nslice_if_pdl to
catch that.  We need to investigate the limitations
of method call syntax.  Maybe a { } block with
the CODE check there could work...

Only a partial solution since a method call can
only be made if the $self is a blessed object.