Add comments with nslice_if_pdl() to findslice
Left these in place but commented out to in case someone
wanted to continue down this development line.  I assumed
that since the modifier syntax is not valid perl that
the only place where nslice would need to be changed
to nslice_if_pdl would be in the filter section for slices
with no modifiers present.