Appear to have working PDL::NiceSlice2
With a workaround to a Filter::Simple bug and a hack
to workaround a missing feature in Text::Balanced, we
now have a trial implementation of an improved
PDL::NiceSlice filter.  To try out, just replace your
exising NiceSlice.pm file with this one.

NOTE: We will still to implement similar processing for
      pdl2/perldl as sourcefilters don't work with evals..