Add support for pdls larger than 2GiB
This commit applies the fix from Jarle Brinchmann
for pdl_grow() and a similar one for pdl_malloc()
replacing the use of int type for nbytes by the
use of STRLEN.  On 64bit platforms, this allows for
pdls larger than 2**31 bytes.

NOTE: The total number of elements in a piddle
is still limited to 2**31.

NOTE: Eventually we need to change the usage of
smalloc (a.k.a. pdl_malloc) and grow (a.k.a. pdl_grow)
to use STRLEN arguments instead.