Removed unnecessary rule for DefaultFlow in PP.pm.
The pp_def key DefaultFlow is only used in two rules. The first
ensures that DefaultFlow has a default value of zero and the
second rule uses that boolean value to set up a code fragment.
However, the first rule can be subsumed by the second rule by
using the optional-condition syntax (an underscore in front of
the condition name). So, I fixed it.