Switch to use Benchmark ':hireswallclock' timings
This allows the total time to be reported with small
enough granularity to see whether there is a performance
improvement from use of pthreads.  It would be nice if
there were a Benchmark variant that was more useful for
showing speedup information.  Thanks to Dima Kogan for
sorting this out and the fix.