Made some PP.pm code more legible.
The XS C Header variable in mkxscat and mkVarArgsxscat was named
chdrs, which made me think "Change Directories", not "C Headers".
So, I renamed it to xs_c_headers for clarity. I also renamed prel
to prelude. It doesn't tell you what kind of prelude it is, but at
least you know it's something that comes before something else,
which I hope helps.

Also, and this may just be a style thing, I revised the code that
constructs the prelude string. It may be just me, but joining a
bunch of members in an array by surrounding the array with quotes
(i.e. $joined = "@seperate") really threw me off. Instead, I used
the join operator, with a suggestive use of the fat comma:

 join $with_this_seperator => (@these_items);

I hope that makes the meaning of the code clearer to anybody else
who comes along.