Moved PMFunc logic from pp_def into a Rule.
The handling of PMFunc was inserted by-hand into pp_def. Given that
all of the optional argument handling is supposed to be done with
Rules, I moved this into its own Rule at the end of the Rule table.
Now we are guaranteed that the key PMFunc will have a meaningful
value if the user did not supply one.