Remove unnecessary HASP2Child key.
The HASP2Child key was essentially a value for P2Child that had the
same boolean evaluation but was guaranteed to be defined as a numeric
value of either 0 or 1. However, the value of HASP2Child was never
used in any arithmetic, only in boolean context. Since undef evaluates
to false, all references to HASP2Child can be replaced by _P2Child
with impunity. I believe that is much clearer. It says, "If the user
indicated a value for P2Child, pass that, otherwise pass the undefined
value." It clearly points back to a user-level key, which I believe
is preferable to an internally defined key derived directly from the
user-level one.