Improved and CPANififed some documentation.
Types.pm is generated from a .pm.PL file which, as is very common,
was never formatted to have its pod directly parsed and displayed.
This means that, when viewed on CPAN, the badvalue documentation
was messed up. I've corrected that.

Furthermore, I updated PP's documentation to reflect that we have
support for a long-long type, with the single-character Q. I also
cleaned up PP's examples that used GenericTypes, which used Perl's
default bareword-to-string interpretation that croaks under strictures.
The modified code makes the processing of GenericTypes clearer, I
hope.