Found a bug with rcols, here's the test.
This commit includes a test which exposes a bug with rcols. Some code
that immediately followed the file-read loop did one last operation,
cleaning up the last line, so to speak. The next block of code checked
if nothing was read in by the file. Problem is, the code that does the
clean-up assumed that some variables had a special structure that
required at least a single pass through the while loop, and an empty
file handle would never go through the loop. So, if you inadvertently
fed an empty file handle to rcols, it would die a sad death saying

Can't use an undefined value as an ARRAY reference at /path/to/PDL/IO/Misc.pm line 676

That's a bug if ever I've seen one, and this test checks for it.