Improved pdl_from_string, more secure, handles BAD string.
I started digging into the just-named function because I thought
it would be handy to add support for things like:

 $a = pdl q[1 2 bad 4]

and have it supply a BAD value instead of zero. I got that working,
but then I realized there are some security holes in the
constructor. I've added some additional checking and barfing for
the most agregious cases, but it's still possible to do damage and
we should look more closely into how to properly perform
pdl_from_string.