PDL::Core::pdl_from_string more secure, handles bad/inf
I've revised how pdl_from_string handles its input. Per the
discussion on the PDL porters list, the only characters it accepts
now are whitespace, numbers, and

 [ ] . , ; + - e E

with three important exceptions. The strings 'inf', 'nan', and 'bad'
all Do What You Mean now. It even goes so far as to check that
your datatype supports inf and croaks if it does not (for integers).
It warns if you try to use nan on an integer datatype and converts
it to a bad value.

This patch also includes a few doc updates and added tests. I also
took the opportunity to organize the test file in a way that is
easier to follow (I hope).