Fixed pdl-from-string for Strawberry Perl.
I discovered that Strawberry Perl (and all Windows?) print inf as
1.#INF. Similar strangeness also occurrs with nan. I've added code
in PDL::Core::new_pdl_from_string to properly process these strings.
Since they're on all platforms, this means that a user can paste
the output of their Windows box on a mailing list, and Linux users
can paste it into their editors and shells and it should work.