Added proper parser error handling for pdl-from-string.
The parser assumed that warnings were on, but they were not. I've
lexically added them to PDL::Core::parse_basic_string.

In replacing the string-eval with the parser, I stopped catching
exceptions, but parse_basic_string throws them. So, I've wrapped
the call to parse_basic_string in an eval *block*, so as to catch
those exceptions.

The error reporting when an exception is caught used to make
reference to an eval. Such a reference is no longer necessary.