Reverted and improved looks_like_number check.
Consider the following code:

 $data = pdl '1.234';

Removing the looks_like_number check to get pdl-from-string to
work cross-platform had a huge (x10) impact on the time it takes to
execute that code. This was deemed unacceptable, so the
looks_like_number check was reinstated. However, a check for the
strings 'nan' and 'inf' was added to short-circuit the logic of
looks_like_number. The current code is about 4.5% slower than the
original code for converting a single numeric string to a piddle,
but it gives the correct results for nan and inf strings on all
platforms.