Add docs for rvals
The implementation of rvals() shows wraparound for integer
data types when the value of r**2 is greater than the capacity
of the datatype.  Also, if a center is not specified, the one
used is rounded down to an exact pixel location.  I've
documented these features in response to sf.net feature
request tracker item 3396761 "fix rvals arithmetic wrap-around
for byte types".  Modifying the implementation to do differently
without costing memory or compatibility was problematic.  At
some point better diagnostics would help but I'm leary of
slowing down fundamental operations for every call to catch
the occaisional oops.  Maybe we could implement a verbose
option that could be conditionally compiled.