Fix PDL::Core::convert return type confusion
This was sf.net feature #3396738, "PDL::Core::convert default return
type" where the type argument for the convert() routine is a
type *number* which caused mysterious results when a string like
"long" was passed in (instead of long(), for example).  atof on
the string gives 0 which is the byte type.  This cleans up the
documentation and adds a barf if the type is not a number.
Later cleanup could include allowing for strings of type names
instead.