Correction to statsover
The ADEV calculation had an erroneous sqrt in the good code, the
bad code, and the docs.  Removed, since the consensus was
that that was plain wrong.  Changed the label in the docs to AADEV,
because what we are really calculating is the _average_ absolute
deviation.  Also reordered PRMS and RMS in the docs, since they
were backwards from the way the function returns them.