Fixed bug in PDL::PP verbose reporting.
First, a section of code was not clearly indented. I fixed that.

Second, PDL::PP::Rule defined a function called report() that is
meant to simplify the reporting of output when verbose output is
requested. (You request verbose output by setting $::PP_VERBOSE to
a true value in your .pd file.) However, it was written as a
function, not an object method; in other words, PDL::PP::Rule
simply called report() with a string to print, like this:

  report('Something useful');

This meant that subclasses of PDL::PP::Rule could not call the
function without explicitly naming the function package.
Unfortunately, they were doing just that. PDL::PP::Rule::Returns
had code like this:

  $self->report('Something useful');

Perl correctly resolved the method to the one defined in
PDL::PP::Rule, but that function simply printed the first argument,
i.e. *the Rule object*. To make matters more confusing, Rules have
a method for stringification, so it was not entirely clear where
this error had crept in. The only real trigger was the lack of
newlines where I expected to see them.

My solution to this problem is to turn report() into a bona fide
method, and to call it as a method wherever it is used. With this
change in place, the verbose reporting finally works correctly.