Fixed two weird variable interpolations in PDL::PP
There was some code to generate text in PDL::PP that did not properly
escape the variables being interpolated. That is, it *improperly*
escaped them, but in a way that worked. Here's an example of the error:

 my $a = 'counter';
 my $string = "  int $a\_int = 0;\n";

When run with warnings, Perl complains that it doesn't know what you
mean by "\_", and it ends up just ignoring the "\" character to
give you this string

 "  int counter_int = 0;\n"

The proper way to do this (which does not produce warnings) is like so:

 my $string = "  int ${a}_int = 0;\n";

I fixed that in a couple of places.