Add connected 4-component labeling to PDL::Image2D.
cc8compt alread existed which provided connected 8-component labeling.
I renamed that to ccNcompt, and added two new PDL subs cc8compt and
cc4compt, which just use ccNcompt as the underlying engine. The
difference between the cc8 and cc4 is just whether the diagonals
are used in the image segmentation.