Add diags for missing PDL::IO::Pic converters
This can cause fails in the various t/pic*.t tests since the
missing converters are not always appropriately checked for
and skipped.  Now the tests will print the actual values
for the converters in case of fails so that we can try to
diagnose the full fix to the problem.