Cleaned up a couple of PDL::PP warnings.
PDL::PP has begun generating warnings when run under the -w flag,
which happens these days under Perl 5.14. I've examined the code
and cleaned it up as best I can. I'm still not sue how the symbol
table class is used, but I'm fairly certain that the setting of
the Sym2Id key in add_ids() is useless since that key is never
used elsewhere. As it is, these changes don't break the compilation
of any of the core modules, so I expect that it didn't do anything
important.