POD cleanup for image2d.pd
We really should revisit the PDL docs generation so that
the docs that are shown on http://search.cpan.org look
right.  Maybe we could add a step in PP that modifies
the docs (e.g., ==head to =head) and so forth so that
the only docs visible would be the generated ones.