Update EU::MM required $VERSION to 6.31
This is the first release of ExtUtils::MakeMaker that has
support for INSTALL_BASE as a location for non-system module
repositories.  It installs in the same fashion as Module::Build's
--install-base option and is *much* easier to comprehend where
things will end up.  I've added a note to DEPENDENCIES that
builds using earlier versions should still be possible with
some minor edits.

We'll need to update the PDL::FAQ related to this.