Use approx in limits_round.t, is_deeply in limits_trans_err.t
It is usually a bad idea to check for floating point
equality as platform differences can cause fails that
aren't significant.  I replaced '==' by approx() in
t/limits_round.t and fixed the implementation of approx()
which was broken for non-piddle inputs.  The method call
syntax should not be needed as PDL::LiteF (the smallest
possible set of modules to run PDL) does include PDL::Ops.

Took a look at t/limits_trans_err.t but couldn't think
of a good way to fix quickly but did notice that
eq_array() has been deprecated in favor of is_deeply()
by Test::More and made that change.