Added commented code, to be uncommented after 2.4.10 is released.
I looked back over the pdl-from-string code and realized that some
of the parsing could be made more strict. For example, at the
moment, the following input:

  $piddle = pdl q[1 +-2 3];

will be interpreted as

  $piddle = pdl q[1 -2 3];

In othe words, the initial plus sign is simply dropped. The new
code includes checks for this condition and croaks if it finds it.