Update VERSION to 2.4.9_992 for RC2 checkout
The focus of RC2 is to ensure that any lingering
code issue surprises (not expected) from RC1 are
addressed/documented and to allow for any POD or
internal doc updates to the source tree.  After
RC2, the remaining work should be on updates to
the web site and on the PDL book.