Improved line number reporting from Core.pm.PL
I had added a # line directive at one place in Core.pm.PL while I
was working on PDL::new_pdl_from_string, but when I encountered
a new error much further along in the file, I figured that the
line reporting was off. It was. This patch adds a number of
additional # line directives to Core.pm.PL so that error reporting
now gives the correct line of the underlying source file.