Fixed a bug in cat's error reporting.
I had code that called PDL::cat(@piddles), but ran into the edge
case that @piddles was completely empty. This particular edge case
was not properly identified by cat's extensive error processing,
so I've added something that correctly diagnoses it.

I also added proper $arg->isa('PDL') checking. The original code
checked if ref($arg) =~ /PDL/, which is not the proper way to
check something is an object of a particular kind. The correct
way to do this is to say

    if (not eval{$arg->isa('PDL')}) ...