Made pp_line_numbers.t a little more robust
I noticed that the number of tests in pp_line_numbers.t was hard
coded, but it actually depended on the number of types the PDL
supports. It now explicitly specifies two types to test (float and
double, though it could be anything).

It also now closes the filehandle to the generated xs file before
removing the file. Previously, extraneous copies of foobar.xs were
left around, at least on Windows machines and maybe on others.