Update PDL delete magic for 64bit perls/OSes
Updated PDL_CORE_VERSION to 9 and replaced int declarations
for lengths and offsets by STRLEN, Size_t, and Off_t as
appropriate.  May need to change some of these a bit but
I believe the size of the types is now correct to support
memory mapping on 64-bit systems and perls.