Fixed PP commentary
PP generates lots of code. Some of that code is confusing at first
blush, so PP added C-style multiline comments to help anybody
examining the generated XS. Unfortunately, if a user had a bug in
their code's logic and attempted to debug by commenting out large
swaths of their PP code, the non-nesting of C-style comments would
invariably lead to weird compile failures.

My proposed solution is to #define the symbol PDL_COMMENT(comment),
thus allowing any commentary such as this:

  _internal_var->value  /* get some internal thing */

to be replaced with something like this:

  _internal_var->value  PDL_COMMENT("get some internal thing")

It's a bit more verbose, but it enables users to use multiline
comments in the way described above.