Preliminary set of 64bit index support modes to Basic/Core
The first stage of the conversion of PDL indexing to support 64bit
platforms is to replace all usages of int, long, and PDL_Long that
are actually indexes into piddles (or the equivalents: dimension
lengths, nvals, ...) by a new PDL_Index datatype that is determined
at build time to correspond to the supported index/pointer sizes for
a given platform.

This first set of conversions is all by inspection and I expect
problems to show up on the first build tests.  Specifically, there
will need to be defined the printf format appropriate to PDL_Index
on a given platform that will need to be used for the built in
diagnostics.

For configuration, the interesting parameters from %Config are:
 - lseeksize and lseektype for file offsets
 - sizesize and sizetype for memory offsets/sizes and indexing

TO CHECK: there are some usages of pdl_packint and pdl_unpackint
in the Core files that maybe should be using the PDL_Index type.