Paired down some barfing/croaking text in PDL::Core
Dima's updates a couple of years ago switched PDL's barf to use a stack trace,
at least in PDL::Core. This becomes a problem for some functions that re-barf.
This minor patch to Core.pm.PL udpates the barfing behavior to use croak where
appropriate.

It also changes the definition of PDL::Core::barf to goto() Carp::confess
instead of simply calling it. The latter behavior resulted in an erroneous line
at the top of the stack trace pointing (uselessly) to the definition of
PDL::Core::barf.