Update VERSION to 2.4.10_006 for final PDL-2.4.11 fixes
This is just a place holder version.  It will be updated
to 2.4.11 when the release distribution is made.  Also
turned off $PDL::debug in t/flexraw.t to quiet test output
and added info to INSTALL and Known_problems regarding the
problem building PDL with some MinGW gcc + win32 perl
configurations due to gcc crashing.