Minor cleanups to pdl_from_string.t, inline-comment-test.t
While working on a few other files in the test suite, I came across
the code in inline-comment-test.t that is supposed to ensure that
the tests are only run on machines that are configured properly to
execute the tests. The code was a bit confusing, so I cleaned it up
in a way that I hope makes it easier for others to follow in the
future.

I also expanded the explanation of the test that is given at the
top of pdl_from_string, since support for inf and nan values is
now one of the goals of the pdl-from-string behavior.