Fixed nan sign issue, as reported by Matthew McGillis
I had noticed that pdl-from-string always said that 'nan' had a
minus sign, but I didn't bother much about it. Matthew McGillis
not only knew better, but wrote a patch to fix the situation,
which is applied here.

The only remaining issue is now the sign of -0, which Matthew
included in his tentative test suite but which does not appear to
work on my machine. Further investigations are needed...