Add note to OPTIMIZE advising the addition of -O0 to -g.
This addresses SF bug #3465663. When I set OPTIMIZE to
'-O0 -g', built and installed PDL, and examined a
segfault-produced core dump in gdb, I get full information
for the PDL functions (including line numbers for pdlcore.c),
but of course not for the Perl functions because my Perl was
not compiled with -g.