Add ENVI file format bailout for PDL::IO::FlexRaw
Since the filename + header file convention is the
same for PDL::IO::FlexRaw files and for ENVI format
files, I've added a check for ENVI and a bailout if
this is not a PDL::IO::FlexRaw format file.

I expect to replace this by a call to readenvi at
some point to smooth things out.