Fix dummy() to generate correct slice() call
Sf.net feature request/bug report #3479009 had a hard to find
problem with dummy() because it was generating an underlying
call to slice() with an invalid string contruct, e.g. '* 128'
rather than '*128'.  This fixes dummy() to coerce the ' 128'
to a number, i.e. '128' so that the slice string is correct as
far as syntax.

Also made one more failing NaN sign test a TODO one since it
was failing on cygwin.