Suppress some of the "Writing MYMETA" messages during 'perl Makefile.PL'.
This should suppress some of the "Writing MYMETA.yml and MYMETA.json" lines
that crop up during the first part of the PDL build.  It will not take care
of all of them, to do that I'll have to hand edit some Makefile.PL's.  Putting
this in now to see if it messes up older versions of EU::MM, and if it's OK
I'll do the rest by hand later.  See pdl-porters thread from 2011-Dec-21.