Removed =for ref Internal pod directives
The pod directive =for ref Internal shows up in many places in
Graphics/PGPLOT/Window/Window.pm, but it does not serve much of a
purpose. It probably served a useful purpose at some point, but
that purpose has been lost. Removing for now, with the expectation
of revisiting after devising a better pdldoc keyword system.