Small tweaks to PLplot Makefile.PL and test file
In the t/plplot.t, added some logic to skip the final test on darwin.
The SVG file PLplot generates is large but valid. Also zero-padded the
names of the output files (test2.svg --> test02.svg) so they would be
in numerical and lexical order in directory listings, for debugging purposes.
In the Makefile.PL, I added some logic to remove the 'temp' executable so
that the build directory would stay clean.